src/backend/test_gateway/requirements.txt
python-socketio
asyncio
aiohttp